pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ha-testcontainer"
version = "0.1.0"
description = "Full test container for Home Assistant with demo config, custom components, and visual testing"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
keywords = ["home-assistant", "testcontainers", "testing", "docker", "playwright"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "testcontainers>=4.9",
    "requests>=2.32",
]

[project.optional-dependencies]
visual = [
    "playwright>=1.50",
    "pytest-playwright>=0.5",
    "Pillow>=10.0",
]
test = [
    "pytest>=8.0",
    "pytest-playwright>=0.5",
    "playwright>=1.50",
    "Pillow>=10.0",
]

[project.urls]
Homepage = "https://github.com/Lint-Free-Technology/ha-testcontainer"
Issues = "https://github.com/Lint-Free-Technology/ha-testcontainer/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
markers = [
    "version_smoke: slow smoke tests that pull stable/beta/dev Docker images (deselected by default)",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ha_testcontainer*"]
